+++ frontend/lendme/lib/services/auth_service.dart
